remove SMTP config
