+++ config_registry.go
@@ -31,7 +31,6 @@ type ConfigSectionRegistry struct {
 
 func resetRegistry() error {
 	ConfigSections := []ConfigSection{
-		&AlertsConfig{},
 		&AmboyConfig{},
 		&APIConfig{},
 		&AuthConfig{},
