+++ rest/model/admin_test.go
@@ -89,9 +89,6 @@ func TestModelConversion(t *testing.T) {
 	}
 	assert.Equal(testSettings.ShutdownWaitSeconds, *apiSettings.ShutdownWaitSeconds)
 
-	assert.EqualValues(testSettings.Alerts.SMTP.From, utility.FromStringPtr(apiSettings.Alerts.SMTP.From))
-	assert.EqualValues(testSettings.Alerts.SMTP.Port, apiSettings.Alerts.SMTP.Port)
-	assert.Equal(len(testSettings.Alerts.SMTP.AdminEmail), len(apiSettings.Alerts.SMTP.AdminEmail))
 	assert.EqualValues(testSettings.Amboy.Name, utility.FromStringPtr(apiSettings.Amboy.Name))
 	assert.EqualValues(testSettings.Amboy.DBConnection.URL, utility.FromStringPtr(apiSettings.Amboy.DBConnection.URL))
 	assert.EqualValues(testSettings.Amboy.DBConnection.Database, utility.FromStringPtr(apiSettings.Amboy.DBConnection.Database))
@@ -153,9 +150,6 @@ func TestModelConversion(t *testing.T) {
 	assert.EqualValues(testSettings.LoggerConfig.Buffer.Count, apiSettings.LoggerConfig.Buffer.Count)
 	assert.EqualValues(testSettings.LoggerConfig.Buffer.UseAsync, apiSettings.LoggerConfig.Buffer.UseAsync)
 	assert.EqualValues(testSettings.LoggerConfig.Buffer.IncomingBufferFactor, apiSettings.LoggerConfig.Buffer.IncomingBufferFactor)
-	assert.EqualValues(testSettings.Notify.SMTP.From, utility.FromStringPtr(apiSettings.Notify.SMTP.From))
-	assert.EqualValues(testSettings.Notify.SMTP.Port, apiSettings.Notify.SMTP.Port)
-	assert.Equal(len(testSettings.Notify.SMTP.AdminEmail), len(apiSettings.Notify.SMTP.AdminEmail))
 	assert.EqualValues(testSettings.Notify.SES.From, utility.FromStringPtr(apiSettings.Notify.SES.From))
 	assert.EqualValues(testSettings.PodLifecycle.S3BaseURL, utility.FromStringPtr(apiSettings.PodLifecycle.S3BaseURL))
 	assert.EqualValues(testSettings.PodLifecycle.MaxParallelPodRequests, apiSettings.PodLifecycle.MaxParallelPodRequests)
@@ -231,9 +225,6 @@ func TestModelConversion(t *testing.T) {
 	dbInterface, err := apiSettings.ToService()
 	assert.NoError(err)
 	dbSettings := dbInterface.(evergreen.Settings)
-	assert.EqualValues(testSettings.Alerts.SMTP.From, dbSettings.Alerts.SMTP.From)
-	assert.EqualValues(testSettings.Alerts.SMTP.Port, dbSettings.Alerts.SMTP.Port)
-	assert.Equal(len(testSettings.Alerts.SMTP.AdminEmail), len(dbSettings.Alerts.SMTP.AdminEmail))
 	assert.EqualValues(testSettings.Amboy.Name, dbSettings.Amboy.Name)
 	assert.EqualValues(testSettings.Amboy.DBConnection, dbSettings.Amboy.DBConnection)
 	assert.EqualValues(testSettings.Amboy.LocalStorage, dbSettings.Amboy.LocalStorage)
@@ -272,9 +263,6 @@ func TestModelConversion(t *testing.T) {
 	assert.EqualValues(testSettings.LoggerConfig.Buffer.Count, dbSettings.LoggerConfig.Buffer.Count)
 	assert.EqualValues(testSettings.LoggerConfig.Buffer.UseAsync, dbSettings.LoggerConfig.Buffer.UseAsync)
 	assert.EqualValues(testSettings.LoggerConfig.Buffer.IncomingBufferFactor, dbSettings.LoggerConfig.Buffer.IncomingBufferFactor)
-	assert.EqualValues(testSettings.Notify.SMTP.From, dbSettings.Notify.SMTP.From)
-	assert.EqualValues(testSettings.Notify.SMTP.Port, dbSettings.Notify.SMTP.Port)
-	assert.Equal(len(testSettings.Notify.SMTP.AdminEmail), len(dbSettings.Notify.SMTP.AdminEmail))
 	assert.EqualValues(testSettings.Notify.SES.From, dbSettings.Notify.SES.From)
 	assert.EqualValues(testSettings.PodLifecycle.S3BaseURL, dbSettings.PodLifecycle.S3BaseURL)
 	assert.EqualValues(testSettings.PodLifecycle.MaxParallelPodRequests, dbSettings.PodLifecycle.MaxParallelPodRequests)
