+++ rest/route/admin_test.go
@@ -143,9 +143,6 @@ func (s *AdminRouteSuite) TestAdminRoute() {
 	settings, ok := settingsResp.(evergreen.Settings)
 	s.True(ok)
 
-	s.EqualValues(testSettings.Alerts.SMTP.From, settings.Alerts.SMTP.From)
-	s.EqualValues(testSettings.Alerts.SMTP.Port, settings.Alerts.SMTP.Port)
-	s.Equal(len(testSettings.Alerts.SMTP.AdminEmail), len(settings.Alerts.SMTP.AdminEmail))
 	s.EqualValues(testSettings.Amboy.Name, settings.Amboy.Name)
 	s.EqualValues(testSettings.Amboy.LocalStorage, settings.Amboy.LocalStorage)
 	s.EqualValues(testSettings.Amboy.GroupDefaultWorkers, settings.Amboy.GroupDefaultWorkers)
@@ -175,9 +172,6 @@ func (s *AdminRouteSuite) TestAdminRoute() {
 	s.EqualValues(testSettings.LoggerConfig.Buffer.Count, settings.LoggerConfig.Buffer.Count)
 	s.EqualValues(testSettings.LoggerConfig.Buffer.UseAsync, settings.LoggerConfig.Buffer.UseAsync)
 	s.EqualValues(testSettings.LoggerConfig.Buffer.IncomingBufferFactor, settings.LoggerConfig.Buffer.IncomingBufferFactor)
-	s.EqualValues(testSettings.Notify.SMTP.From, settings.Notify.SMTP.From)
-	s.EqualValues(testSettings.Notify.SMTP.Port, settings.Notify.SMTP.Port)
-	s.Equal(len(testSettings.Notify.SMTP.AdminEmail), len(settings.Notify.SMTP.AdminEmail))
 	s.EqualValues(testSettings.PodLifecycle.S3BaseURL, settings.PodLifecycle.S3BaseURL)
 	s.EqualValues(testSettings.PodLifecycle.MaxParallelPodRequests, settings.PodLifecycle.MaxParallelPodRequests)
 	s.EqualValues(testSettings.PodLifecycle.MaxPodDefinitionCleanupRate, settings.PodLifecycle.MaxPodDefinitionCleanupRate)
